+++ client/src/pages/Recipe/AddPost.tsx
@@ -57,13 +57,8 @@ const AddPost = () => {
   );
   const [tagsDatas, setTagsDatas] = useState<ITagsData[]>([]);
 
-  // // 빈 값 체크
-  // const isJsonDataEmpty = useRecipeJsonDataValidation(
-  //   data,
-  //   ingredientsDatas,
-  //   directDatas,
-  //   tagsDatas
-  // );
+  // 빈 값 체크
+  const isJsonDataEmpty = useRecipeJsonDataValidation(recipeData.inputTexts);
 
   const inputHandler = (
     e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
@@ -72,93 +67,94 @@ const AddPost = () => {
     dispatch(recipeActions.setTitleOrBody({ [target.name]: target.value }));
   };
 
-  // const submitHandler = async (event: FormEvent) => {
-  //   event.preventDefault();
+  const submitHandler = async (event: FormEvent) => {
+    event.preventDefault();
+    console.log("submit 이벤트");
 
-  //   /** 텍스트 누락 체크 */
-  //   if (isJsonDataEmpty === true) {
-  //     message.fire({
-  //       icon: "error",
-  //       title:
-  //         "레시피 등록에 실패했습니다.\n 누락된 정보가 있는지 \n확인해주세요.",
-  //     });
-  //     return;
-  //   }
+    /** 텍스트 누락 체크 */
+    if (isJsonDataEmpty === true) {
+      message.fire({
+        icon: "error",
+        title:
+          "레시피 등록에 실패했습니다.\n 누락된 정보가 있는지 \n확인해주세요.",
+      });
+      return;
+    }
 
-  //   /** 이미지 누락 체크 */
-  //   const emptyImageIndex = stepImgFiles.findIndex((el) => el === undefined);
-  //   if (emptyImageIndex >= 0) {
-  //     message.fire({
-  //       icon: "error",
-  //       title: `요리 순서의 ${emptyImageIndex + 1}번째 이미지를 \n추가해주세요`,
-  //     });
-  //     return;
-  //   }
-  //   if (!thumbNail || stepImgFiles.length === 0) {
-  //     message.fire({
-  //       icon: "error",
-  //       title: `등록하려면 \n이미지를 추가해주세요.`,
-  //     });
-  //     return;
-  //   }
+    //   /** 이미지 누락 체크 */
+    //   const emptyImageIndex = stepImgFiles.findIndex((el) => el === undefined);
+    //   if (emptyImageIndex >= 0) {
+    //     message.fire({
+    //       icon: "error",
+    //       title: `요리 순서의 ${emptyImageIndex + 1}번째 이미지를 \n추가해주세요`,
+    //     });
+    //     return;
+    //   }
+    //   if (!thumbNail || stepImgFiles.length === 0) {
+    //     message.fire({
+    //       icon: "error",
+    //       title: `등록하려면 \n이미지를 추가해주세요.`,
+    //     });
+    //     return;
+    //   }
 
-  //   /** 서버 요청 데이터 구축 */
-  //   const formData = new FormData();
-  //   formData.append("imgThumbNail", thumbNail);
+    //   /** 서버 요청 데이터 구축 */
+    //   const formData = new FormData();
+    //   formData.append("imgThumbNail", thumbNail);
 
-  //   stepImgFiles.forEach((file) => {
-  //     if (file) {
-  //       formData.append("imgDirection", file);
-  //     }
-  //   });
+    //   stepImgFiles.forEach((file) => {
+    //     if (file) {
+    //       formData.append("imgDirection", file);
+    //     }
+    //   });
 
-  //   // 재료순서 변경 인덱스 정렬
-  //   const filteredIngredients: TypeOfIngredients[] = [];
-  //   ingredientsDatas.forEach((oneline, idx) => {
-  //     filteredIngredients.push({ ...oneline, index: idx + 1 });
-  //   });
+    //   // 재료순서 변경 인덱스 정렬
+    //   const filteredIngredients: TypeOfIngredients[] = [];
+    //   ingredientsDatas.forEach((oneline, idx) => {
+    //     filteredIngredients.push({ ...oneline, index: idx + 1 });
+    //   });
 
-  //   // 조리순서 변경 인덱스 정렬
-  //   const filteredDirects: IStepValues[] = [];
-  //   directDatas.forEach((oneline, idx) => {
-  //     filteredDirects.push({ ...oneline, index: idx + 1 });
-  //   });
+    //   // 조리순서 변경 인덱스 정렬
+    //   const filteredDirects: IStepValues[] = [];
+    //   directDatas.forEach((oneline, idx) => {
+    //     filteredDirects.push({ ...oneline, index: idx + 1 });
+    //   });
 
-  //   const recipeDatas = {
-  //     ...data,
-  //     category: checkedCateg,
-  //     ingredients: filteredIngredients,
-  //     directions: filteredDirects,
-  //     tags: tagsDatas,
-  //   };
+    //   const recipeDatas = {
+    //     ...data,
+    //     category: checkedCateg,
+    //     ingredients: filteredIngredients,
+    //     directions: filteredDirects,
+    //     tags: tagsDatas,
+    //   };
 
-  //   formData.append(
-  //     "recipe",
-  //     new Blob([JSON.stringify(recipeDatas)], { type: "application/json" })
-  //   );
+    //   formData.append(
+    //     "recipe",
+    //     new Blob([JSON.stringify(recipeDatas)], { type: "application/json" })
+    //   );
 
-  //   /** 서버 요청 */
-  //   const response = await axios.post("/api/v1/recipe/add", formData, {
-  //     headers: { "content-type": "multipart/form-data" },
-  //   });
+    //   /** 서버 요청 */
+    //   const response = await axios.post("/api/v1/recipe/add", formData, {
+    //     headers: { "content-type": "multipart/form-data" },
+    //   });
 
-  //   try {
-  //     const newId = response.data.data.id;
-  //     navigate(`/post/${newId}/`);
-  //   } catch (error) {
-  //     message.fire({
-  //       icon: "error",
-  //       title:
-  //         "레시피 등록에 실패했습니다.\n 누락된 정보가 있는지 \n확인해주세요.",
-  //     });
-  //   }
-  // };
+    //   try {
+    //     const newId = response.data.data.id;
+    //     navigate(`/post/${newId}/`);
+    //   } catch (error) {
+    //     message.fire({
+    //       icon: "error",
+    //       title:
+    //         "레시피 등록에 실패했습니다.\n 누락된 정보가 있는지 \n확인해주세요.",
+    //     });
+    //   }
+  };
 
   return (
     <SFormContainer>
       <SLogoRecipe src={recipeLogo} alt="recipeLogo"></SLogoRecipe>
-      <form action="" method="post">
-        {/* <form action="" method="post" onSubmit={submitHandler}> */}
+      {/* <form action="" method="post"> */}
+      <form action="" method="post" onSubmit={submitHandler}>
         <SSection>
           <SRecipeInfo>
             <SRecipeTexts>
@@ -231,9 +227,9 @@ const AddPost = () => {
           <SFormBtn color={"var(--deep-green)"} type="reset">
             취소
           </SFormBtn>
-          {/* <SFormBtn type="button" onClick={submitHandler}>
+          <SFormBtn type="button" onClick={submitHandler}>
             등록
-          </SFormBtn> */}
+          </SFormBtn>
           {/* <button >임시저장</button> */}
         </SSectionBtn>
       </form>
